refactor: 상품 BC에 따른 패키지 구조 리팩터링
