+++ src/main/java/kitchenpos/menus/application/MenuService.java
@@ -5,8 +5,8 @@ import kitchenpos.menus.domain.model.MenuGroup;
 import kitchenpos.menus.domain.model.MenuGroupRepository;
 import kitchenpos.menus.domain.model.MenuProduct;
 import kitchenpos.menus.domain.model.MenuRepository;
-import kitchenpos.domain.Product;
-import kitchenpos.domain.ProductRepository;
+import kitchenpos.products.model.Product;
+import kitchenpos.products.model.ProductRepository;
 import kitchenpos.shared.external.PurgomalumClient;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
