+++ src/main/java/kitchenpos/products/infrastructure/JpaProductRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.domain;
+package kitchenpos.products.infrastructure;
 
+import kitchenpos.products.model.Product;
+import kitchenpos.products.model.ProductRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
