+++ src/main/java/kitchenpos/products/model/Product.java
@@ -1,9 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.products.model;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
 
 import java.math.BigDecimal;
 import java.util.UUID;
