+++ src/test/java/kitchenpos/Fixtures.java
@@ -8,7 +8,7 @@ import kitchenpos.domain.OrderLineItem;
 import kitchenpos.domain.OrderStatus;
 import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.OrderType;
-import kitchenpos.domain.Product;
+import kitchenpos.products.model.Product;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
