+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -4,8 +4,8 @@ import kitchenpos.menus.domain.model.Menu;
 import kitchenpos.menus.domain.model.MenuGroupRepository;
 import kitchenpos.menus.domain.model.MenuProduct;
 import kitchenpos.menus.domain.model.MenuRepository;
-import kitchenpos.domain.Product;
-import kitchenpos.domain.ProductRepository;
+import kitchenpos.products.model.Product;
+import kitchenpos.products.model.ProductRepository;
 import kitchenpos.shared.external.PurgomalumClient;
 import kitchenpos.menus.application.MenuService;
 import org.junit.jupiter.api.BeforeEach;
